define ma57int to size_t instead of long, so it also works on 32bit systems (suggested by J. Currie)

+++ Ipopt/src/Algorithm/LinearSolvers/IpMa57TSolverInterface.hpp
@@ -13,7 +13,7 @@
 #include "IpSparseSymLinearSolverInterface.hpp"
 
 #ifdef FUNNY_MA57_FINT
-typedef long ma57int;
+typedef size_t ma57int;
 #else
 typedef ipfint ma57int;
 #endif
